changes related to razorpay amount and total

+++ src/main/java/com/quma/quma_shopify_backend/services/PaymentService.java
@@ -78,7 +78,7 @@ public class PaymentService {
             String receipt = order.getOrderId();
 
             com.razorpay.Order rpOrder = razorpayService.createOrder(
-                    order.getAmount().longValue(),
+                    order.getAmount(),
                     receipt);
 
             // Save razorpay order details
